Show a progress bar for plugin installation during pulumi up
